+++ src/columns/HyPerCol.hpp
@@ -113,10 +113,10 @@ public:
    int numCommRows()                      {return icComm->numCommRows();}
 
    // a random seed based on column id
-   unsigned long getSeed() { return random_seed; }
-   unsigned long getObjectSeed(int count) { unsigned long seed = random_seed_obj; random_seed_obj += count; return seed;}
+   unsigned int getSeed() { return random_seed; }
+   unsigned int getObjectSeed(int count) { unsigned long seed = random_seed_obj; random_seed_obj += count; return seed;}
 
-   unsigned long getRandomSeed();
+   unsigned int getRandomSeed();
       // Nov. 28, 2012.  All MPI processes get the same base seed, and should use global information to seed individual neurons.
       // {return (unsigned long) time((time_t *) NULL); } // Aug 21, 2012: Division by 1+columnId() moved to calling routine}
 
@@ -236,11 +236,11 @@ private:
                                    // Value 2: layers have form NameOfLayer.pvp
    int filenamesContainConnectionNames; // Similar to filenamesContainLayerNames, but for connections
 
-   unsigned long random_seed;
-   unsigned long random_seed_obj;  // Objects that need to generate random numbers should request a seed from
-                                   // the HyPerCol, saying how many they need (across all processes in an MPI run).
-                                   // random_seed_obj is incremented by the number requested, so that everything
-                                   // that needs a random seed gets a unique seed, and things are reproducible.
+   unsigned int random_seed;
+   unsigned int random_seed_obj;  // Objects that need to generate random numbers should request a seed from
+                                  // the HyPerCol, saying how many they need (across all processes in an MPI run).
+                                  // random_seed_obj is incremented by the number requested, so that everything
+                                  // that needs a random seed gets a unique seed, and things are reproducible.
 
 }; // class HyPerCol
